requirements.txt
crewai>=0.28.8
crewai-tools>=0.1.6
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.6.1
python-multipart>=0.0.6
requests>=2.31.0
python-dotenv>=1.0.0
aiofiles>=23.2.1